third_party/blink/web_tests/virtual/webnn-service-on-cpu/external/wpt/webnn/idlharness.https.any.worker-expected.txt

This is a testharness.js-based test.
Found 3 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] MLGraphBuilder interface: operation argMin(MLOperand, optional MLArgMinMaxOptions)
  assert_equals: property has wrong .length expected 1 but got 2
[FAIL] MLGraphBuilder interface: operation argMax(MLOperand, optional MLArgMinMaxOptions)
  assert_equals: property has wrong .length expected 1 but got 2
[FAIL] MLGraphBuilder interface: operation softmax(MLOperand, unsigned long)
  assert_equals: property has wrong .length expected 2 but got 1
Harness: the test ran to completion.
